cleaned up the typescript wrapper code
