--- packages/coln-js-runtime/src/ts/runtime.ts
@@ -1,83 +1,0 @@
-import { Value, RowView, tryValueRowId } from "./value";
-import { toTxnValue, valueEqual } from "./value";
-import { StorageCtx } from "./store";
-
-export type Tuple = Value[];
-
-export namespace Tuple {
-  export function equal(t0: readonly Value[], t1: readonly Value[]): boolean {
-    if (t0.length == t1.length) {
-      for (var i = 0; i < t0.length; i++) {
-        if (!valueEqual(t0[i], t1[i])) {
-          return false;
-        }
-      }
-      return true;
-    } else {
-      return false;
-    }
-  }
-}
-
-export interface ReadonlySet {
-  has(x: Value): boolean;
-  values(): Iterator<RowView>;
-}
-
-export interface ReadWriteSet extends ReadonlySet {
-  add(): Value;
-}
-
-export class RelTable<W> {
-  path: string[];
-  ctx: StorageCtx<W>;
-
-  constructor(path: string[], ctx: StorageCtx<W>) {
-    this.path = path;
-    this.ctx = ctx;
-  }
-
-  apply_to(params: Value[]): AppliedRelTable<W> {
-    return new AppliedRelTable(this, params);
-  }
-
-  storePath(): string {
-    return this.path.join(".");
-  }
-}
-
-export class AppliedRelTable<W> implements ReadWriteSet {
-  relation: RelTable<W>;
-  params: Value[];
-
-  constructor(relation: RelTable<W>, params: Value[]) {
-    this.relation = relation;
-    this.params = params;
-  }
-
-  has(x: Value): boolean {
-    if (x.tag !== "row_id" && x.tag !== "row_handle") return false;
-
-    const rowId = tryValueRowId(x);
-    if (rowId === undefined) return false;
-
-    const row = this.relation.ctx.rowById(this.relation.storePath(), rowId);
-
-    return row !== undefined && Tuple.equal(row.values, this.params);
-  }
-
-  values(): Iterator<RowView> {
-    const rows = this.relation.ctx.scanTable(this.relation.storePath());
-
-    return rows.filter((row) => Tuple.equal(row.values, this.params)).values();
-  }
-
-  add(): Value {
-    const handle = this.relation.ctx.add(
-      this.relation.storePath(),
-      this.params.map(toTxnValue),
-    );
-
-    return { tag: "row_handle", value: handle };
-  }
-}
