--- packages/coln-js-runtime/src/ts/schema.ts
@@ -1,9 +1,0 @@
-// TODO interface as determined by the compiler
-
-export interface Schema<T> {
-  get_json(): string;
-}
-
-export interface SchemaProvider<T> {
-  schema(): Schema<T>;
-}
