+++ packages/coln-js-runtime/src/ts/set.ts
@@ -1,0 +1,11 @@
+import { Value } from "./value.ts"
+import { RowView  } from "#wasm-bodge/bindings";
+
+export interface View {
+  has(x: Value): boolean;
+  values(): Iterator<RowView>;
+}
+
+export interface Transaction extends View {
+  add(): Value;
+}
